N900PVPUDNH7 4.4.4
Conflicts:

	Makefile
	arch/arm/boot/dts/msm8974/msm8974-sec.dtsi
	arch/arm/mach-msm/mpm-of.c
	arch/arm/mm/fault.c
	crypto/api.c
	drivers/cpufreq/cpufreq.c
	drivers/cpufreq/cpufreq_interactive.c
	drivers/cpufreq/cpufreq_ondemand.c
	drivers/gpu/msm/kgsl_iommu.c
	drivers/net/usb/asix.c
	drivers/platform/msm/qpnp-power-on.c
	drivers/staging/android/ashmem.c
	drivers/video/msm/mdss/Kconfig
	drivers/video/msm/mdss/mdnie_lite_tuning.c
	drivers/video/msm/mdss/mdss_mdp_intf_cmd.c
	drivers/video/msm/mdss/mdss_samsung_dsi_panel.c
	fs/ext4/ialloc.c
	fs/ext4/xattr.c
	init/main.c
	kernel/exit.c
	kernel/time/timekeeping.c
	mm/page-writeback.c
	mm/page_alloc.c
	net/wireless/nl80211.c
	security/selinux/include/objsec.h

+++ drivers/net/wireless/bcmdhd4354/bcmsdh_linux.c
@@ -1,7 +1,7 @@
 /*
  * SDIO access interface for drivers - linux specific (pci only)
  *
- * Copyright (C) 1999-2013, Broadcom Corporation
+ * Copyright (C) 1999-2014, Broadcom Corporation
  * 
  *      Unless you and Broadcom execute a separate written software license
  * agreement governing use of this software, this software is licensed to you
@@ -21,7 +21,7 @@
  * software in any way with any other Broadcom software provided under a license
  * other than the GPL, without Broadcom's express prior written consent.
  *
- * $Id: bcmsdh_linux.c 444405 2013-12-19 12:28:07Z $
+ * $Id: bcmsdh_linux.c 461443 2014-03-12 02:40:59Z $
  */
 
 /**
@@ -44,6 +44,7 @@ extern void dhdsdio_isr(void * args);
 #include <bcmutils.h>
 #include <dngl_stats.h>
 #include <dhd.h>
+
 #include <dhd_linux.h>
 
 /* driver info, initialized when bcmsdh_register is called */
@@ -133,7 +134,7 @@ bcmsdh_chipmatch(uint16 vendor, uint16 device)
 void* bcmsdh_probe(osl_t *osh, void *dev, void *sdioh, void *adapter_info, uint bus_type,
 	uint bus_num, uint slot_num)
 {
-	uint32 regs;
+	ulong regs;
 	bcmsdh_info_t *bcmsdh;
 	uint32 vendevid;
 	bcmsdh_os_info_t *bcmsdh_osinfo = NULL;
